Build the runtime with warnings on, and act on what they found
Only the decoder was compiled with -Wall -Wextra; the runtime and the lifter
were not, so nothing had ever checked them. They are now on for every
hand-written source. The recompiled output is deliberately excluded -- it is
machine-emitted and would bury the signal, which is preferable to leaving the
whole target unchecked, since that is exactly how this went unnoticed.

The one finding that matters: imports.c still held a set of joystick handlers
that report no devices and JOYERR_UNPLUGGED. Nothing registers them, because
gamepad.c took that surface over when hotplug was implemented -- but if they
were ever reconnected they would silently disable the controller feature that
is one of this project's reasons to exist. Deleted rather than left as a
tombstone.

hostwin_present took a palette parameter that has been dead since the 8-bit
surface assumption was dropped; every caller passed NULL, and the pixel
parameter was still named "indexed" though surfaces are XRGB. Both corrected,
so the signature stops implying palettes are still in play.

The rest were genuinely trivial: three signed/unsigned comparisons against
enum bounds, and eight unused stub helpers, all removed.

Verified after: ctest 3/3, and a full match still shows 11,473 keyed blits and
8 sound effects with music playing.

+++ runtime/gdi.c
@@ -147,7 +147,7 @@ static Bitmap *bmp_load(const char *path)
 
 static uint32_t bitmap_handle(Bitmap *b)
 {
-    if (!b || nbitmaps >= MAX_GDI) return 0;
+    if (!b || nbitmaps >= (int)MAX_GDI) return 0;
     bitmaps[nbitmaps] = b;
     return H_BITMAP + (uint32_t)nbitmaps++;
 }
@@ -288,7 +288,7 @@ static void h_LoadImageA(void)
 
 static void h_CreateCompatibleDC(void)
 {
-    if (ndcs >= MAX_GDI) { ret_stdcall(1, 0); return; }
+    if (ndcs >= (int)MAX_GDI) { ret_stdcall(1, 0); return; }
     dc_bitmap[ndcs] = 0;
     ret_stdcall(1, H_DC + (uint32_t)ndcs++);
 }
